refactor(coding): shrink tui composition root
